Struct sui_security_watchdog::SecurityWatchdogConfig

pub struct SecurityWatchdogConfig {
    pub pd_wallet_monitoring_service_id: String,
    pub config: PathBuf,
    pub sf_account_identifier: Option<String>,
    pub sf_warehouse: Option<String>,
    pub sf_database: Option<String>,
    pub sf_schema: Option<String>,
    pub sf_username: Option<String>,
    pub sf_role: Option<String>,
    pub client_metric_host: String,
    pub client_metric_port: u16,
}

Fields
pd_wallet_monitoring_service_id: Stringconfig: PathBufsf_account_identifier: Option<String>sf_warehouse: Option<String>sf_database: Option<String>sf_schema: Option<String>sf_username: Option<String>sf_role: Option<String>client_metric_host: String

The url of the metrics client to connect to.

client_metric_port: u16

The port of the metrics client to connect to.